Split dbconnect into submodules
